+++ src/app/components/fav-list/fav-list.component.html
@@ -32,9 +32,6 @@
                     <th
                         class="px-6 align-middle border border-solid py-3 text-xs uppercase border-l-0 border-r-0 whitespace-nowrap font-bold text-left">
                         Detalles</th>
-                    <th
-                        class="px-6 align-middle border border-solid py-3 text-xs uppercase border-l-0 border-r-0 whitespace-nowrap font-bold text-left">
-                        Eliminar</th>
                 </tr>
             </thead>
 
@@ -54,10 +51,6 @@
                     <td
                         class="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4 font-semibold">
                         <button>Ver</button>
-                    </td>
-                    <td
-                        class="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4 font-semibold">
-                        <button>Borrar</button>
                     </td>
                 </tr>
             </tbody>
