--- src/app/components/list/create-list/create-list.component.html
@@ -1,35 +1,0 @@
-<div class="container mx-auto px-4 h-full" *ngIf="login">
-    <div class="flex content-center items-center justify-center h-full">
-        <div class="w-full lg:w-4/12 px-4">
-            <div
-                class="relative flex flex-col min-w-0 break-words w-full mb-6 shadow-lg rounded-lg bg-blueGray-200 border-0">
-                <div class="rounded-t mb-0 px-6 py-6"></div>
-                <div class="flex-auto px-4 lg:px-10 py-10 pt-0">
-                    <form>
-                        <div class="relative w-full mb-3">
-                            <label class="block uppercase text-blueGray-600 text-xs font-bold mb-2"
-                                htmlFor="grid-password" for="name">Nombre</label>
-                            <input type="text"
-                                class="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
-                                placeholder="Nombre de la lista" id="name" name="name" />
-                        </div>
-
-                        <div class="relative w-full mb-3">
-                            <label class="block uppercase text-blueGray-600 text-xs font-bold mb-2"
-                                htmlFor="grid-password" for="description">Descripción</label>
-                            <textarea
-                                class="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
-                                placeholder="Descripción de la lista" id="description" name="description"></textarea>
-                        </div>
-
-                        <div class="text-center mt-6">
-                            <input
-                                class="bg-blueGray-800 text-white active:bg-blueGray-600 text-sm font-bold uppercase px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 w-full"
-                                type="submit" [routerLink]="['/auth/create-list']" value="Crear lista" />
-                        </div>
-                    </form>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
