+++ src/app/components/sidebar/sidebar.component.html
@@ -119,13 +119,13 @@
 
         <li class="items-center">
           <a [routerLink]="['/admin/films']" class="text-xs uppercase py-3 font-bold block" routerLinkActive
-            #adminPerson="routerLinkActive" [ngClass]="
-            adminPerson.isActive
+            #adminFilm="routerLinkActive" [ngClass]="
+            adminFilm.isActive
                 ? 'text-red-600 hover:text-red-700'
                 : 'text-blueGray-700 hover:text-blueGray-500'
             ">
             <i class="fas fa-solid fa-users mr-2 text-sm" [ngClass]="
-            adminPerson.isActive ? 'opacity-75' : 'text-blueGray-300'
+            adminFilm.isActive ? 'opacity-75' : 'text-blueGray-300'
               "></i>
             Películas
           </a>
