--- src/app/models/dto/create-list.dto.ts
@@ -1,11 +1,0 @@
-export class CreateListDto {
-    name: string = '';
-    description: string = '';
-    language: string = '';
-
-    constructor(name: string, description: string, language: string) {
-        this.name = name;
-        this.description = description;
-        this.language = language;
-    }
-}
