+++ src/app/models/interfaces/favorite/get-list.interface.ts
