--- src/app/models/interfaces/list/list-details-dialog.interface.ts
@@ -1,5 +1,0 @@
-import { ListDetailsResponse } from "./list-details.interface";
-
-export interface DialogDataList {
-    listInfo: ListDetailsResponse;
-}
