--- src/app/models/interfaces/list/list-details.interface.ts
@@ -1,31 +1,0 @@
-// Generated by https://quicktype.io
-
-export interface ListDetailsResponse {
-    created_by: string;
-    description: string;
-    favorite_count: number;
-    id: string;
-    items: Item[];
-    item_count: number;
-    iso_639_1: string;
-    name: string;
-    poster_path: string;
-}
-
-export interface Item {
-    poster_path: string;
-    adult: boolean;
-    overview: string;
-    release_date: string;
-    original_title: string;
-    genre_ids: number[];
-    id: number;
-    media_type: string;
-    original_language: string;
-    title: string;
-    backdrop_path: null;
-    popularity: number;
-    vote_count: number;
-    video: boolean;
-    vote_average: number;
-}
