--- src/app/models/interfaces/list/new-list.interface.ts
@@ -1,6 +1,0 @@
-export interface NewListsResponse {
-    status_message: string;
-    success: boolean;
-    stastus_code: number;
-    list_id: number;
-}
