--- src/app/models/interfaces/person/person-details-dialog.interface.ts
@@ -1,7 +1,0 @@
-import { PersonDetailsResponse } from "./person-details.interface";
-import { KnownFor } from "./person.interface";
-
-export interface DialogDataPerson {
-    personInfo: PersonDetailsResponse;
-    movies: KnownFor[];
-}
