+++ src/app/views/auth/favorite/favorite.component.html
@@ -1,0 +1,1 @@
+<app-favorite-list></app-favorite-list>
