--- src/app/views/auth/list/account-list/account-list.component.html
@@ -1,1 +1,0 @@
-<app-show-list></app-show-list>
