--- src/app/views/auth/list/new-list/new-list.component.html
@@ -1,1 +1,0 @@
-<app-create-list></app-create-list>
