--- src/app/views/auth/list/new-list/new-list.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-new-list',
-  templateUrl: './new-list.component.html',
-  styleUrls: ['./new-list.component.css']
-})
-export class NewListComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
